move snackbar creation outside service
